+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## **4.1.0** (in progress)
 
 ### Features added:
-* Add required field indication and validation to invoice field mapping profile (UIDATIMP-877)
+* Add required field indication and validation to invoice field mapping profile (UIDATIMP-877)UIDATIMP-885
+* Numeric subfield mappings are not working. Fixed (UIDATIMP-885)
 
 ## [4.0.1](https://github.com/folio-org/ui-data-import/tree/v4.0.1) (2021-04-02)