+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Cover `<CodeHighlight>` component with tests (UIDATIMP-702)
 * Cover `<FolioRecordTypeSelect>` component with tests (UIDATIMP-705)
 * Cover `<JobLogsContainer>` component with tests (UIDATIMP-713)
+* Cover`<ListTemplate>` component with tests (UIDATIMP-707)
 
 ## [4.0.0](https://github.com/folio-org/ui-data-import/tree/v4.0.0) (2021-03-18)