+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change history for ui-data-import
 
 ## **4.1.0** (in progress)
+
+### Features added:
+* Add required field indication and validation to invoice field mapping profile (UIDATIMP-877)
+* Numeric subfield mappings are not working. Fixed (UIDATIMP-885)
+
+## [4.0.1](https://github.com/folio-org/ui-data-import/tree/v4.0.1) (2021-04-02)
+
 ### Features added:
 * Adjust the UI for action profiles when linked to job profiles (UIDATIMP-749)
 * Adjust the UI for action profiles when linked to field mapping profiles (UIDATIMP-870)
@@ -8,6 +15,11 @@
 * MARC Updates field mapping profile: be able to collapse/expand accordions (UIDATIMP-858)
 * Ensure file name info for records posted direct via API on landing page (UIDATIMP-873)
 * Log lite: Import job summary for EDIFACT invoices (UIDATIMP-816)
+* Cover `<CodeHighlight>` component with tests (UIDATIMP-702)
+* Cover `<FolioRecordTypeSelect>` component with tests (UIDATIMP-705)
+* Cover `<JobLogsContainer>` component with tests (UIDATIMP-713)
+* Cover`<ListTemplate>` component with tests (UIDATIMP-707)
+* Log lite - Invoice JSON screen (UIDATIMP-817)
 
 ## [4.0.0](https://github.com/folio-org/ui-data-import/tree/v4.0.0) (2021-03-18)
 
@@ -61,8 +73,6 @@
 * Refactor Action profiles to use final-form instead of redux-form (UIDATIMP-827)
 * Add personal data disclosure form (UIDATIMP-786)
 * Adjust "vendor name" field value (UIDATIMP-868)
-* Cover `<FolioRecordTypeSelect>` component with tests (UIDATIMP-705)
-* Cover `<CodeHighlight>` component with tests (UIDATIMP-702)
 
 ### Bugs fixed:
 * Fix Accessibility problems for settings/data-import/match-profiles (lists must only directly contain li elements) (UIDATIMP-452)